Formatting edit grant
