+++ frontend/src/main-page/grants/grant-view/GrantView.tsx
@@ -12,20 +12,17 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import ContactCard from "./ContactCard";
 import GrantFieldCol from "./GrantFieldCol";
 import { CostBenefitAnalysis } from "./CostBenefitAnalysis";
-import EditGrant from "../filter-bar/EditGrant";
+import EditGrant from "../edit-grant/EditGrant";
 
 interface GrantItemProps {
   grant: Grant;
 }
 
 const GrantItem: React.FC<GrantItemProps> = observer(({ grant }) => {
-
   const [showEditGrant, setShowEditGrant] = useState(false);
-  
-  
-  useEffect(() => {
-  }, [grant]);
-  
+
+  useEffect(() => {}, [grant]);
+
   const useTruncatedElement = ({
     ref,
   }: {
@@ -60,11 +57,7 @@ const GrantItem: React.FC<GrantItemProps> = observer(({ grant }) => {
 
   function formatDate(isoString: string): string {
     if (!isoString) return "N/A";
-    const date = new Date(isoString);
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-    const year = date.getFullYear();
-    return `${month}/${day}/${year}`;
+    return new Date(isoString + "T00:00:00").toLocaleDateString();
   }
 
   function formatCurrency(amount: number): string {
@@ -74,205 +67,205 @@ const GrantItem: React.FC<GrantItemProps> = observer(({ grant }) => {
 
   return (
     <div>
-    <div className="w-full bg-white rounded-md flex flex-col gap-6 p-6">
-      {/* Top header part */}
-      <div className="flex justify-between">
-        {/* Left side */}
-        <div className="flex flex-col gap-2 items-start text-left">
-          <h2 className="text-2xl font-semibold">{grant.organization}</h2>
-          <StatusIndicator curStatus={grant.status} />
+      <div className="w-full bg-white rounded-md flex flex-col gap-6 p-6">
+        {/* Top header part */}
+        <div className="flex justify-between">
+          {/* Left side */}
+          <div className="flex flex-col gap-2 items-start text-left">
+            <h2 className="text-2xl font-semibold">{grant.organization}</h2>
+            <StatusIndicator curStatus={grant.status} />
+          </div>
+          {/* Right side */}
+          <div className="flex flex-col gap-2 items-end">
+            <Button
+              text="Edit"
+              onClick={() => setShowEditGrant(true)}
+              className="bg-white text-black border-2 border-grey-500"
+              logo={faPenToSquare}
+              logoPosition="right"
+            />
+          </div>
         </div>
-        {/* Right side */}
-        <div className="flex flex-col gap-2 items-end">
-          <Button
-            text="Edit"
-            onClick={() => setShowEditGrant(true)}
-            className="bg-white text-black border-2 border-grey-500"
-            logo={faPenToSquare}
-            logoPosition="right"
-          />
-        </div>
-      </div>
-      <hr className="border-grey-400 border-t-2 rounded-full" />
+        <hr className="border-grey-400 border-t-2 rounded-full" />
 
-      {/* Middle info part */}
-      <div className="flex flex-col gap-6 items-start text-left">
-        {/* Description */}
-        <GrantFieldCol
-          fields={[
-            {
-              label: "Description",
-              item: (
-                <div>
-                  <p
-                    ref={ref}
-                    className={` ${!isShowingMore && "line-clamp-3"}`}
-                    onClick={toggleIsShowingMore}
-                  >
-                    {grant?.description || "N/A"}
-                  </p>
-                  {isTruncated && (
-                    <button
-                      className="text-secondary"
-                      onClick={toggleIsShowingMore}
-                    >
-                      {isShowingMore ? "Show less" : "show more"}
-                    </button>
-                  )}
-                </div>
-              ),
-            },
-          ]}
-        />
-        {/* Other details */}
-        <div className="grid grid-cols-2 lg:grid-cols-3 xl:grid-cols-5 w-full gap-4">
+        {/* Middle info part */}
+        <div className="flex flex-col gap-6 items-start text-left">
+          {/* Description */}
           <GrantFieldCol
             fields={[
               {
-                label: "Amount ($)",
-                value: formatCurrency(grant.amount),
-                important: true,
-              },
-              {
-                label: "BCAN Eligible",
-                important: true,
-                item: grant.does_bcan_qualify ? (
-                  <span className="text-green">
-                    <FontAwesomeIcon icon={faCheckSquare} /> Yes
-                  </span>
-                ) : (
-                  <span className="text-red">
-                    <FontAwesomeIcon icon={faXmarkSquare} /> No
-                  </span>
+                label: "Description",
+                item: (
+                  <div>
+                    <p
+                      ref={ref}
+                      className={` ${!isShowingMore && "line-clamp-3"}`}
+                      onClick={toggleIsShowingMore}
+                    >
+                      {grant?.description || "N/A"}
+                    </p>
+                    {isTruncated && (
+                      <button
+                        className="text-secondary"
+                        onClick={toggleIsShowingMore}
+                      >
+                        {isShowingMore ? "Show less" : "show more"}
+                      </button>
+                    )}
+                  </div>
                 ),
               },
             ]}
           />
+          {/* Other details */}
+          <div className="grid grid-cols-2 lg:grid-cols-3 xl:grid-cols-5 w-full gap-4">
+            <GrantFieldCol
+              fields={[
+                {
+                  label: "Amount ($)",
+                  value: formatCurrency(grant.amount),
+                  important: true,
+                },
+                {
+                  label: "BCAN Eligible",
+                  important: true,
+                  item: grant.does_bcan_qualify ? (
+                    <span className="text-green">
+                      <FontAwesomeIcon icon={faCheckSquare} /> Yes
+                    </span>
+                  ) : (
+                    <span className="text-red">
+                      <FontAwesomeIcon icon={faXmarkSquare} /> No
+                    </span>
+                  ),
+                },
+              ]}
+            />
+            <GrantFieldCol
+              fields={[
+                {
+                  label: "Due Date",
+                  important: true,
+                  value: formatDate(grant.application_deadline),
+                },
+                {
+                  label: "Application Date",
+                  value: "TBD (add to DB)",
+                },
+              ]}
+            />
+            <GrantFieldCol
+              fields={[
+                {
+                  label: "Grant Start Date",
+                  value: formatDate(grant.grant_start_date),
+                },
+                {
+                  label: "Report Deadlines",
+                  item:
+                    grant.report_deadlines &&
+                    grant.report_deadlines.length > 0 ? (
+                      grant.report_deadlines.map(
+                        (deadline: string, index: number) => (
+                          <div key={index} className="text-black">
+                            {formatDate(deadline)}
+                          </div>
+                        ),
+                      )
+                    ) : (
+                      <div className="">N/A</div>
+                    ),
+                },
+              ]}
+            />
+            <GrantFieldCol
+              colspan={2}
+              fields={[
+                {
+                  label: "Timeline (years)",
+                  value: grant.timeline,
+                },
+                {
+                  label: "Estimated Completion Time (hours)",
+                  value: grant.estimated_completion_time,
+                },
+              ]}
+            />
+          </div>
+        </div>
+        <hr className="border-grey-400 border-t-2 rounded-full" />
+
+        {/* Bottom info */}
+        <div className="flex flex-col gap-6 items-start text-left">
+          {/* Contacts */}
           <GrantFieldCol
             fields={[
               {
-                label: "Due Date",
-                important: true,
-                value: formatDate(grant.application_deadline),
-              },
-              {
-                label: "Application Date",
-                value: "TBD (add to DB)",
+                label: "Contacts",
+                item: (
+                  <div className="grid grid-cols-1 xl:grid-cols-2 w-full h-full lg:w-[85%] gap-6">
+                    <ContactCard contact={grant.bcan_poc} type="BCAN" />
+                    <ContactCard
+                      contact={grant.grantmaker_poc}
+                      type="Granter"
+                    />
+                  </div>
+                ),
               },
             ]}
           />
+          {/* Documents */}
           <GrantFieldCol
             fields={[
               {
-                label: "Grant Start Date",
-                value: formatDate(grant.grant_start_date),
-              },
-              {
-                label: "Report Deadlines",
+                label: "Documents",
                 item:
-                  grant.report_deadlines &&
-                  grant.report_deadlines.length > 0 ? (
-                    grant.report_deadlines.map(
-                      (deadline: string, index: number) => (
-                        <div key={index} className="text-black">
-                          {formatDate(deadline)}
-                        </div>
-                      ),
-                    )
+                  grant.attachments && grant.attachments.length > 0 ? (
+                    <div className="columns-2 xl:columns-4 gap-4 lg:w-[90%]">
+                      {grant.attachments.map((attachment, index) => (
+                        <p key={index} className="truncate w-full mb-1">
+                          <a
+                            href={attachment.url}
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            className="text-secondary underline"
+                          >
+                            {attachment.attachment_name || attachment.url}
+                          </a>
+                        </p>
+                      ))}
+                    </div>
                   ) : (
                     <div className="">N/A</div>
                   ),
               },
             ]}
           />
+        </div>
+        <hr className="border-grey-400 border-t-2 rounded-full" />
+
+        {/* Cost Benefit */}
+        <div className="flex flex-col gap-6 items-start text-left">
           <GrantFieldCol
-            colspan={2}
             fields={[
               {
-                label: "Timeline (years)",
-                value: grant.timeline,
-              },
-              {
-                label: "Estimated Completion Time (hours)",
-                value: grant.estimated_completion_time,
+                label: "Cost Analysis Calculator",
+                item: <CostBenefitAnalysis grant={grant} />,
               },
             ]}
           />
         </div>
       </div>
-      <hr className="border-grey-400 border-t-2 rounded-full" />
-
-      {/* Bottom info */}
-      <div className="flex flex-col gap-6 items-start text-left">
-        {/* Contacts */}
-        <GrantFieldCol
-          fields={[
-            {
-              label: "Contacts",
-              item: (
-                <div className="grid grid-cols-1 xl:grid-cols-2 w-full h-full lg:w-[85%] gap-6">
-                  <ContactCard contact={grant.bcan_poc} type="BCAN" />
-                  <ContactCard contact={grant.grantmaker_poc} type="Granter" />
-                </div>
-              ),
-            },
-          ]}
-        />
-        {/* Documents */}
-        <GrantFieldCol
-          fields={[
-            {
-              label: "Documents",
-              item:
-                grant.attachments && grant.attachments.length > 0 ? (
-                  <div className="columns-2 xl:columns-4 gap-4 lg:w-[90%]">
-                    {grant.attachments.map((attachment, index) => (
-                      <p key={index} className="truncate w-full mb-1">
-                        <a
-                          href={attachment.url}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-secondary underline"
-                        >
-                          {attachment.attachment_name || attachment.url}
-                        </a>
-                      </p>
-                    ))}
-                  </div>
-                ) : (
-                  <div className="">N/A</div>
-                ),
-            },
-          ]}
-        />
+      <div>
+        {showEditGrant && (
+          <EditGrant
+            grantToEdit={grant}
+            onClose={async () => {
+              setShowEditGrant(false);
+            }}
+          />
+        )}
       </div>
-      <hr className="border-grey-400 border-t-2 rounded-full" />
-
-      {/* Cost Benefit */}
-      <div className="flex flex-col gap-6 items-start text-left">
-      <GrantFieldCol
-          fields={[
-            {
-              label: "Cost Analysis Calculator",
-              item: (
-                <CostBenefitAnalysis grant={grant} />
-              ),
-            },
-          ]}
-        />
-    </div>
-    </div>
-    <div>
-    {showEditGrant && (
-            <EditGrant
-              grantToEdit={grant}
-              onClose={async () => {
-                setShowEditGrant(false);
-              }}
-              isOpen={showEditGrant}
-            />
-          )}
-          </div>
     </div>
   );
 });
